fix(migrate): Improve and doc widget infra setup
Added JLv3-compatible setup to the workshop and demo templates, and
expanded docs in notebook and README on what extra setup may be
needed for users running in their own environments.

It turns out that ipycanvas does need setup in SMStudio/NBI JL3 too,
and the setup is much easier to do via lifecycle config rather than
trying to do manually and restart the server.